requirements.txt
python-telegram-bot==12.8
python-telegram-bot-pagination
requests
tpblite
